Replace 'priority' terminology with 'weight' in task handling
The concept of 'Priority' in task handling has been refactored to 'Weight' across the entire project. This includes changes in function and variable names as well as comments and README. This change is to better reflect the actual nature of task importance determination as not just a priority queue but handling based on a quantitive 'weight' value.